requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
streamlit>=1.22.0
pandas>=1.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0